Plugins: Fix loading of backend plugins
